fix(disk-analyzer): unregister worktrees on delete and split modules
Delete now runs git worktree remove for linked checkouts, and Atmos
workspaces (including archived-on-disk) go through workspace delete so
the DB stays consistent. Drop bare dist/build/out/tmp/output cleanup
hints. Split the engine and service disk_analyzer files into focused
modules.
