+++ apps/web/src/features/disk-analyzer/lib/tree-adapters.ts
@@ -471,14 +471,9 @@ const CLEANUP_HINTS: Record<string, string> = {
   ".vitest": "Vitest cache",
   ".swc": "SWC compiler cache",
   ".nx": "Nx computation cache",
-  dist: "Build distribution output",
-  build: "Build output (Gradle/CMake/web/etc.)",
-  out: "Compile/export output",
-  output: "Generic build output",
   ".cache": "Tool cache directory",
   ".tmp": "Temporary build files",
   ".temp": "Temporary build files",
-  tmp: "Temporary files (project-local)",
   target: "Rust/Cargo or sbt/Scala build artifacts",
   __pycache__: "Python bytecode cache",
   ".pytest_cache": "pytest cache",
