+++ crates/core-service/src/service/disk_analyzer/overview.rs
@@ -1,138 +1,40 @@
-//! Disk Analyzer business service — scan sessions, ownership, project roots.
-//!
-//! Default scans cover Atmos-related paths (`~/.atmos`, imported projects,
-//! Atmos.app) plus linked git worktrees and code-agent **session**
-//! directories (not whole agent homes). The first overview paint is
-//! Atmos + session dirs; leftover worktrees are grouped in a second wave.
-//! Opt-in `scan_all` walks the user home (+ Applications) and still badges
-//! those paths.
+//! Atmos overview assembly and the two-wave overview walk.
 
 use std::collections::{HashMap, HashSet};
 use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicBool, AtomicU64, AtomicUsize, Ordering};
 use std::sync::Arc;
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
 use core_engine::{
-    agent_data_roots, clear_path_cache, finalize_tree, invalidate_path_cache,
-    node_needs_wider_children, CleanupSuggestion, DiskAnalyzerEngine, DiskNode, DiskPathKind,
-    DiskScanRoots, DiskVolumeInfo, FsEngine, GitEngine, ProgressCallback, ScanProgress, ScanStats,
-    ScanStatus, DEFAULT_TREE_DEPTH,
+    agent_data_roots, finalize_tree, CleanupSuggestion, DiskAnalyzerEngine, DiskNode, DiskPathKind,
+    DiskScanRoots, GitEngine, ProgressCallback, ScanProgress, ScanStats, ScanStatus,
 };
 use parking_lot::Mutex;
-use serde::Serialize;
 use serde_json::{json, Value};
 use tokio::sync::broadcast;
-use uuid::Uuid;
 
-use crate::service::{project::ProjectService, workspace::WorkspaceService};
+use crate::service::project::ProjectService;
+use crate::service::workspace::WorkspaceService;
 use crate::{Result, ServiceError};
 
-const MAX_SESSIONS: usize = 8;
-const SESSION_TTL: Duration = Duration::from_secs(30 * 60);
-const EVENT_CHANNEL_CAPACITY: usize = 64;
-const DEFAULT_MAX_CHILDREN: usize = 30;
-/// Key used in the partial-entry map for the user home tree (scan-all mode).
-const ENTRY_KEY_HOME: &str = "home";
-/// Key used in the partial-entry map for `/Applications` (scan-all mode).
-const ENTRY_KEY_APPLICATIONS: &str = "applications";
-/// Synthetic overview root for Atmos-scoped multi-entry scans.
-const ATMOS_OVERVIEW_PATH: &str = "atmos://disk-usage";
-const ATMOS_OVERVIEW_NAME: &str = "Atmos";
-const AGENT_GROUP_PATH: &str = "atmos://disk-usage/agent-data";
-const AGENT_GROUP_NAME: &str = "Agent data";
-const WORKTREE_GROUP_PATH: &str = "atmos://disk-usage/git-worktrees";
-const WORKTREE_GROUP_NAME: &str = "Git worktrees";
-/// Cap concurrent `du` during overview (each is I/O heavy; unlimited thrashing made
-/// large roots like `~/.atmos` finish last and take minutes).
-const OVERVIEW_DU_CONCURRENCY: usize = 3;
-/// Keep overview → entry → entry-children so Mole-like artifacts
-/// (`target`, `node_modules`, `.next`, `workspaces`, …) stay visible without
-/// an extra drill before anything useful appears.
-const OVERVIEW_TREE_DEPTH: usize = 3;
+use super::{
+    path_key, AtmosEntryKind, DiskAnalyzerScanEvent, DiskAnalyzerService, DiskAnalyzerSession,
+    EntryRoot, AGENT_GROUP_NAME, AGENT_GROUP_PATH, ATMOS_OVERVIEW_NAME, ATMOS_OVERVIEW_PATH,
+    ENTRY_KEY_APPLICATIONS, ENTRY_KEY_HOME, OVERVIEW_DU_CONCURRENCY, OVERVIEW_TREE_DEPTH,
+    WORKTREE_GROUP_NAME, WORKTREE_GROUP_PATH,
+};
 
-#[derive(Debug, Clone, Serialize)]
-pub struct DiskAnalyzerScanEvent {
-    pub owner_conn_id: String,
-    pub payload: Value,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[allow(dead_code)]
-enum AtmosEntryKind {
-    Project,
-    Workspace,
-    GitWorktree,
-    AgentData,
-}
-
-#[derive(Clone)]
-struct EntryRoot {
-    /// Display label under the overview root (e.g. Home, Applications, project name).
-    label: String,
-    path: PathBuf,
-    /// Set when this entry is an identified Atmos project or workspace root.
-    kind: Option<AtmosEntryKind>,
-}
-
-struct DiskAnalyzerSession {
-    owner_conn_id: String,
-    cancel: Arc<AtomicBool>,
-    tree: Option<Arc<DiskNode>>,
-    inflight_root: Option<String>,
-    stats: Option<ScanStats>,
-    suggestions: Option<Vec<CleanupSuggestion>>,
-    /// Session logical root (overview synthetic path, or a single real path).
-    root_path: PathBuf,
-    /// Real filesystem roots allowed for delete / get_tree (home, Applications, …).
-    entry_roots: Vec<PathBuf>,
-    max_children: usize,
-    /// Identified Atmos project roots (main repo paths).
-    project_roots: Vec<PathBuf>,
-    /// Identified Atmos workspace worktree paths.
-    workspace_roots: Vec<PathBuf>,
-    /// Linked git worktrees discovered on this machine (not Atmos workspaces).
-    git_worktree_roots: Vec<PathBuf>,
-    /// Code-agent session / transcript directories.
-    agent_data_roots: Vec<PathBuf>,
-    /// True after home discovery has populated worktree / agent roots (may be empty).
-    marks_ready: bool,
-    started_at: Instant,
-    completed_at: Option<Instant>,
-    #[allow(dead_code)]
-    is_overview: bool,
-}
-
-pub struct DiskAnalyzerService {
-    engine: DiskAnalyzerEngine,
-    fs_engine: FsEngine,
-    project_service: Arc<ProjectService>,
-    workspace_service: Arc<WorkspaceService>,
-    sessions: Arc<Mutex<HashMap<String, DiskAnalyzerSession>>>,
-    event_tx: broadcast::Sender<DiskAnalyzerScanEvent>,
+#[derive(Debug, Clone, Copy)]
+pub(super) enum OverviewKind {
+    /// Home tree with `/Applications` grafted under home.
+    HomeWithApps,
+    /// Synthetic "Atmos" root; each entry is a top-level child.
+    AtmosSynthetic,
 }
 
 impl DiskAnalyzerService {
-    pub fn new(
-        project_service: Arc<ProjectService>,
-        workspace_service: Arc<WorkspaceService>,
-    ) -> Self {
-        let (event_tx, _) = broadcast::channel(EVENT_CHANNEL_CAPACITY);
-        Self {
-            engine: DiskAnalyzerEngine::new(),
-            fs_engine: FsEngine::new(),
-            project_service,
-            workspace_service,
-            sessions: Arc::new(Mutex::new(HashMap::new())),
-            event_tx,
-        }
-    }
-
-    pub fn subscribe_events(&self) -> broadcast::Receiver<DiskAnalyzerScanEvent> {
-        self.event_tx.subscribe()
-    }
-
-    fn user_home(&self) -> Result<PathBuf> {
+    pub(super) fn user_home(&self) -> Result<PathBuf> {
         let home = self.fs_engine.get_home_dir()?;
         let home = std::fs::canonicalize(&home).unwrap_or(home);
         // Must be a real user home like /Users/<name> or /home/<name> — never `/`.
@@ -163,7 +65,7 @@ impl DiskAnalyzerService {
     }
 
     /// Full-home scan: home is the display root; Applications is grafted under it.
-    fn home_entry_roots(&self) -> Result<Vec<EntryRoot>> {
+    pub(super) fn home_entry_roots(&self) -> Result<Vec<EntryRoot>> {
         let mut entries = Vec::new();
         let home = self.user_home()?;
         tracing::info!(home = %home.display(), "disk analyzer scan-all root is user home");
@@ -186,7 +88,7 @@ impl DiskAnalyzerService {
     }
 
     /// Atmos-scoped roots: `~/.atmos`, app bundle, app support, plus imported projects.
-    fn atmos_static_entry_roots(&self) -> Result<Vec<EntryRoot>> {
+    pub(super) fn atmos_static_entry_roots(&self) -> Result<Vec<EntryRoot>> {
         let home = self.user_home()?;
         let mut entries = Vec::new();
         let mut push_unique = |label: String, path: PathBuf| {
@@ -242,7 +144,7 @@ impl DiskAnalyzerService {
     /// Add identified project roots that are **not** already under a static Atmos entry
     /// (e.g. not under `~/.atmos`). Workspaces live under `~/.atmos/workspaces` and are
     /// tagged on drill-in instead of listed twice.
-    fn push_identified_project_entries(
+    pub(super) fn push_identified_project_entries(
         entries: &mut Vec<EntryRoot>,
         projects: &[(String, PathBuf)],
     ) {
@@ -279,7 +181,7 @@ impl DiskAnalyzerService {
     }
 
     #[allow(dead_code)]
-    fn scan_roots(
+    pub(super) fn scan_roots(
         project_roots: &[PathBuf],
         workspace_roots: &[PathBuf],
         git_worktree_roots: &[PathBuf],
@@ -293,7 +195,7 @@ impl DiskAnalyzerService {
         }
     }
 
-    fn kind_for_entry(
+    pub(super) fn kind_for_entry(
         kind: Option<AtmosEntryKind>,
         path: &str,
         roots: &DiskScanRoots,
@@ -319,14 +221,14 @@ impl DiskAnalyzerService {
         }
     }
 
-    fn apply_kind(node: &mut DiskNode, kind: DiskPathKind) {
+    pub(super) fn apply_kind(node: &mut DiskNode, kind: DiskPathKind) {
         node.is_project = kind.is_project;
         node.is_workspace = kind.is_workspace;
         node.is_git_worktree = kind.is_git_worktree;
         node.is_agent_data = kind.is_agent_data;
     }
 
-    fn dir_shell(name: String, path: String, kind: DiskPathKind) -> DiskNode {
+    pub(super) fn dir_shell(name: String, path: String, kind: DiskPathKind) -> DiskNode {
         DiskNode {
             name,
             path,
@@ -343,7 +245,7 @@ impl DiskAnalyzerService {
         }
     }
 
-    fn overview_entry_labels(entries: &[EntryRoot]) -> HashMap<String, String> {
+    pub(super) fn overview_entry_labels(entries: &[EntryRoot]) -> HashMap<String, String> {
         entries
             .iter()
             .map(|e| {
@@ -357,7 +259,7 @@ impl DiskAnalyzerService {
             .collect()
     }
 
-    fn path_covered_by_entries(path: &Path, entries: &[EntryRoot]) -> bool {
+    pub(super) fn path_covered_by_entries(path: &Path, entries: &[EntryRoot]) -> bool {
         let canon = std::fs::canonicalize(path).unwrap_or_else(|_| path.to_path_buf());
         entries.iter().any(|e| {
             let er = std::fs::canonicalize(&e.path).unwrap_or_else(|_| e.path.clone());
@@ -365,7 +267,7 @@ impl DiskAnalyzerService {
         })
     }
 
-    fn unique_entry_label(base: String, entries: &[EntryRoot]) -> String {
+    pub(super) fn unique_entry_label(base: String, entries: &[EntryRoot]) -> String {
         if !entries.iter().any(|e| e.label == base) {
             return base;
         }
@@ -378,7 +280,7 @@ impl DiskAnalyzerService {
         format!("{base}-extra")
     }
 
-    fn worktree_overview_label(path: &Path) -> String {
+    pub(super) fn worktree_overview_label(path: &Path) -> String {
         let name = path
             .file_name()
             .map(|s| s.to_string_lossy().into_owned())
@@ -400,7 +302,7 @@ impl DiskAnalyzerService {
         }
     }
 
-    fn append_discovered_overview_entries(
+    pub(super) fn append_discovered_overview_entries(
         entries: &mut Vec<EntryRoot>,
         git_worktrees: &[PathBuf],
         agent_roots: &[(String, PathBuf)],
@@ -429,413 +331,9 @@ impl DiskAnalyzerService {
         }
     }
 
-    pub async fn start_scan(
-        &self,
-        owner_conn_id: &str,
-        path: Option<&str>,
-        max_children: Option<usize>,
-        scan_all: bool,
-    ) -> Result<Value> {
-        let max_children = max_children.unwrap_or(DEFAULT_MAX_CHILDREN).max(1);
-        let scan_id = Uuid::new_v4().to_string();
-        let cancel = Arc::new(AtomicBool::new(false));
-
-        // Collect identified Atmos projects + workspaces for tagging / overview entries.
-        let (project_roots, workspace_roots, project_labels) =
-            Self::collect_atmos_locations_static(
-                Arc::clone(&self.project_service),
-                Arc::clone(&self.workspace_service),
-                GitEngine::new(),
-            )
-            .await;
-
-        let (is_overview, overview_kind, root_path, entry_specs) = match path {
-            Some(p) if !p.is_empty() => {
-                let mut root = self.fs_engine.expand_path(p)?;
-                if root.parent().is_none() {
-                    let home = self.user_home()?;
-                    tracing::warn!(
-                        "disk analyzer refused filesystem root; using home {}",
-                        home.display()
-                    );
-                    root = home;
-                }
-                let label = root
-                    .file_name()
-                    .map(|s| s.to_string_lossy().to_string())
-                    .unwrap_or_else(|| root.to_string_lossy().to_string());
-                (
-                    false,
-                    OverviewKind::HomeWithApps,
-                    root.clone(),
-                    vec![EntryRoot {
-                        label,
-                        path: root,
-                        kind: None,
-                    }],
-                )
-            }
-            _ if scan_all => {
-                let entries = self.home_entry_roots()?;
-                if entries.is_empty() {
-                    return Err(ServiceError::Validation(
-                        "No scannable entry paths found".into(),
-                    ));
-                }
-                let home = entries
-                    .iter()
-                    .find(|e| e.label == ENTRY_KEY_HOME)
-                    .map(|e| e.path.clone())
-                    .ok_or_else(|| ServiceError::Validation("Home path missing".into()))?;
-                (true, OverviewKind::HomeWithApps, home, entries)
-            }
-            _ => {
-                let mut entries = self.atmos_static_entry_roots()?;
-                Self::push_identified_project_entries(&mut entries, &project_labels);
-                if entries.is_empty() {
-                    return Err(ServiceError::Validation(
-                        "No Atmos paths found to scan (expected ~/.atmos or imported projects)"
-                            .into(),
-                    ));
-                }
-                tracing::info!(
-                    entries = entries.len(),
-                    projects = project_roots.len(),
-                    workspaces = workspace_roots.len(),
-                    "disk analyzer default scan is Atmos-scoped"
-                );
-                (
-                    true,
-                    OverviewKind::AtmosSynthetic,
-                    PathBuf::from(ATMOS_OVERVIEW_PATH),
-                    entries,
-                )
-            }
-        };
-
-        let entry_roots: Vec<PathBuf> = entry_specs.iter().map(|e| e.path.clone()).collect();
-
-        // Keep the 3-day path cache across scans. Entries revalidate via directory
-        // mtime (any change → miss). Clearing here made every Rescan/auto-start cold
-        // and re-ran multi-minute `du` on ~/.atmos every time.
-        // Cache is still cleared after delete (see delete_path).
-
-        {
-            let mut sessions = self.sessions.lock();
-            // Rescan / scope switch: drop this connection's previous sessions so the
-            // MAX_SESSIONS cap cannot trap the user after Cancel/Rescan loops.
-            Self::cancel_and_purge_owner(&mut sessions, owner_conn_id);
-            Self::evict_expired(&mut sessions);
-            if sessions.len() >= MAX_SESSIONS {
-                Self::evict_oldest_completed(&mut sessions);
-            }
-            // Last resort: drop oldest in-flight sessions (already cancelled above for this owner).
-            while sessions.len() >= MAX_SESSIONS {
-                if !Self::evict_oldest_any(&mut sessions) {
-                    break;
-                }
-            }
-            if sessions.len() >= MAX_SESSIONS {
-                return Err(ServiceError::Validation(
-                    "Too many active disk analyzer scans; cancel one or wait for TTL cleanup"
-                        .into(),
-                ));
-            }
-            sessions.insert(
-                scan_id.clone(),
-                DiskAnalyzerSession {
-                    owner_conn_id: owner_conn_id.to_string(),
-                    cancel: Arc::clone(&cancel),
-                    tree: None,
-                    inflight_root: Some(path_key(&root_path)),
-                    stats: None,
-                    suggestions: None,
-                    root_path: root_path.clone(),
-                    entry_roots: entry_roots.clone(),
-                    max_children,
-                    project_roots: project_roots.clone(),
-                    workspace_roots: workspace_roots.clone(),
-                    git_worktree_roots: vec![],
-                    agent_data_roots: vec![],
-                    marks_ready: false,
-                    started_at: Instant::now(),
-                    completed_at: None,
-                    is_overview,
-                },
-            );
-        }
-
-        let owner = owner_conn_id.to_string();
-        let scan_id_task = scan_id.clone();
-        let root_path_for_task = root_path.clone();
-        let sessions = Arc::clone(&self.sessions);
-        let event_tx = self.event_tx.clone();
-
-        tokio::spawn(async move {
-            let root_path = root_path_for_task;
-            tokio::task::yield_now().await;
-
-            if is_overview {
-                Self::spawn_overview_walk(
-                    sessions,
-                    event_tx,
-                    scan_id_task,
-                    owner,
-                    entry_specs,
-                    project_roots,
-                    workspace_roots,
-                    max_children,
-                    cancel,
-                    overview_kind,
-                );
-            } else {
-                let single = entry_specs
-                    .into_iter()
-                    .next()
-                    .map(|e| e.path)
-                    .unwrap_or_else(|| root_path.clone());
-                Self::spawn_walk(
-                    sessions,
-                    event_tx,
-                    scan_id_task,
-                    owner,
-                    single,
-                    DiskScanRoots {
-                        project_roots,
-                        workspace_roots,
-                        git_worktree_roots: vec![],
-                        agent_data_roots: vec![],
-                    },
-                    max_children,
-                    cancel,
-                    true,
-                );
-            }
-        });
-
-        Ok(json!({
-            "scan_id": scan_id,
-            "root_path": root_path.to_string_lossy(),
-            "status": "started",
-            "overview": is_overview,
-            "scan_all": scan_all,
-        }))
-    }
-
-    pub fn cancel_scan(&self, owner_conn_id: &str, scan_id: &str) -> Result<Value> {
-        let mut sessions = self.sessions.lock();
-        let session = sessions
-            .get(scan_id)
-            .ok_or_else(|| ServiceError::NotFound(format!("scan {scan_id}")))?;
-        Self::ensure_owner(session, owner_conn_id)?;
-        // Signal the walk to stop, then free the session slot immediately.
-        session.cancel.store(true, Ordering::Relaxed);
-        sessions.remove(scan_id);
-        Ok(json!({ "ok": true, "scan_id": scan_id }))
-    }
-
-    pub fn get_tree(
-        &self,
-        owner_conn_id: &str,
-        scan_id: &str,
-        path: Option<&str>,
-        max_children: Option<usize>,
-    ) -> Result<Value> {
-        let (target_key, scan_roots, session_max, cancel, entry_roots, inflight) = {
-            let sessions = self.sessions.lock();
-            let session = sessions
-                .get(scan_id)
-                .ok_or_else(|| ServiceError::NotFound(format!("scan {scan_id}")))?;
-            Self::ensure_owner(session, owner_conn_id)?;
-
-            let target_key = match path {
-                Some(p)
-                    if !p.is_empty() && (p == ATMOS_OVERVIEW_PATH || p.starts_with("atmos://")) =>
-                {
-                    // Synthetic Atmos overview root — not a real filesystem path.
-                    p.to_string()
-                }
-                Some(p) if !p.is_empty() => {
-                    let expanded = self.fs_engine.expand_path(p)?;
-                    Self::ensure_under_any(&expanded, &session.entry_roots)?;
-                    path_key(&expanded)
-                }
-                _ => path_key(&session.root_path),
-            };
-
-            if let Some(tree) = session.tree.as_ref() {
-                if let Some(node) = find_node(tree.as_ref(), &target_key) {
-                    // Truncated leaves need a fresh walk; only return cache when expanded.
-                    if !node.is_dir || node.children_loaded {
-                        let max = max_children.unwrap_or(session.max_children).max(1);
-                        let synthetic =
-                            target_key == ATMOS_OVERVIEW_PATH || target_key.starts_with("atmos://");
-                        // Pruned snapshots cannot grow past the original cap. Re-walk
-                        // when the UI asks for more than the cached real children.
-                        let serve_cache =
-                            !node.is_dir || synthetic || !node_needs_wider_children(node, max);
-                        if serve_cache {
-                            let mut cloned = node.clone();
-                            finalize_tree(&mut cloned, max, DEFAULT_TREE_DEPTH, true);
-                            return Ok(json!({
-                                "status": "ready",
-                                "tree": cloned,
-                                "stats": session.stats.clone(),
-                            }));
-                        }
-                    }
-                }
-            }
-
-            // Only skip spawn when this exact path is already being scanned.
-            // Having session.tree (overview cache) must NOT block drill-in expansion —
-            // overview entries are often measure-only shells with children_loaded=false.
-            let inflight = session
-                .inflight_root
-                .as_ref()
-                .map(|p| p == &target_key)
-                .unwrap_or(false);
-            (
-                target_key,
-                DiskScanRoots {
-                    project_roots: session.project_roots.clone(),
-                    workspace_roots: session.workspace_roots.clone(),
-                    git_worktree_roots: session.git_worktree_roots.clone(),
-                    agent_data_roots: session.agent_data_roots.clone(),
-                },
-                session.max_children,
-                Arc::clone(&session.cancel),
-                session.entry_roots.clone(),
-                inflight,
-            )
-        };
-
-        if inflight {
-            return Ok(json!({
-                "status": "loading",
-                "path": target_key,
-                "stats": Value::Null,
-            }));
-        }
-
-        // Synthetic overview roots cannot be walk()'d on disk.
-        if target_key == ATMOS_OVERVIEW_PATH || target_key.starts_with("atmos://") {
-            return Ok(json!({
-                "status": "loading",
-                "path": target_key,
-                "stats": Value::Null,
-            }));
-        }
-
-        let target_path = PathBuf::from(&target_key);
-        Self::ensure_under_any(&target_path, &entry_roots)?;
-        let max = max_children.unwrap_or(session_max).max(1);
-        {
-            let mut sessions = self.sessions.lock();
-            if let Some(session) = sessions.get_mut(scan_id) {
-                session.inflight_root = Some(target_key.clone());
-                if max > session.max_children {
-                    session.max_children = max;
-                }
-            }
-        }
-
-        Self::spawn_walk(
-            Arc::clone(&self.sessions),
-            self.event_tx.clone(),
-            scan_id.to_string(),
-            owner_conn_id.to_string(),
-            target_path,
-            scan_roots,
-            max,
-            cancel,
-            false,
-        );
-
-        Ok(json!({
-            "status": "loading",
-            "path": target_key,
-            "stats": Value::Null,
-        }))
-    }
-
-    pub async fn delete_path(
-        &self,
-        owner_conn_id: &str,
-        scan_id: &str,
-        path: &str,
-        permanent: bool,
-    ) -> Result<Value> {
-        let (entry_roots, _) = {
-            let sessions = self.sessions.lock();
-            let session = sessions
-                .get(scan_id)
-                .ok_or_else(|| ServiceError::NotFound(format!("scan {scan_id}")))?;
-            Self::ensure_owner(session, owner_conn_id)?;
-            (session.entry_roots.clone(), session.root_path.clone())
-        };
-
-        if path.starts_with("atmos://") {
-            return Err(ServiceError::Validation(
-                "cannot delete a synthetic disk-analyzer path".into(),
-            ));
-        }
-
-        let path = self.fs_engine.expand_path(path)?;
-        let allowed_root = entry_roots
-            .iter()
-            .find(|root| path == **root || path.starts_with(root))
-            .cloned()
-            .ok_or_else(|| {
-                ServiceError::Validation(format!(
-                    "path {} is outside scan entry roots",
-                    path.display()
-                ))
-            })?;
-
-        let engine = DiskAnalyzerEngine::new();
-        let path_for_task = path.clone();
-
-        let freed = tokio::task::spawn_blocking(move || {
-            engine.delete_path(&path_for_task, permanent, Some(&allowed_root))
-        })
-        .await
-        .map_err(|e| ServiceError::Processing(format!("delete task join failed: {e}")))??;
-
-        // Drop caches for deleted path and its parent so sizes refresh on next open.
-        invalidate_path_cache(&path);
-        if let Some(parent) = path.parent() {
-            invalidate_path_cache(parent);
-        }
-        clear_path_cache(); // simple + correct: force full remeasure after delete
-
-        {
-            let mut sessions = self.sessions.lock();
-            if let Some(session) = sessions.get_mut(scan_id) {
-                session.tree = None;
-                session.completed_at = None;
-            }
-        }
-
-        Ok(json!({
-            "success": true,
-            "path": path.to_string_lossy(),
-            "freed_bytes": freed,
-            "permanent": permanent,
-        }))
-    }
-
-    pub fn disk_info(&self, path: Option<&str>) -> Result<DiskVolumeInfo> {
-        let path = match path {
-            Some(p) if !p.is_empty() => self.fs_engine.expand_path(p)?,
-            _ => self.user_home()?,
-        };
-        Ok(self.engine.disk_info(&path)?)
-    }
-
     /// Walk entry roots in parallel and merge into one overview tree.
     #[allow(clippy::too_many_arguments)]
-    fn spawn_overview_walk(
+    pub(super) fn spawn_overview_walk(
         sessions: Arc<Mutex<HashMap<String, DiskAnalyzerSession>>>,
         event_tx: broadcast::Sender<DiskAnalyzerScanEvent>,
         scan_id: String,
@@ -1564,168 +1062,9 @@ impl DiskAnalyzerService {
         });
     }
 
-    #[allow(clippy::too_many_arguments)]
-    fn spawn_walk(
-        sessions: Arc<Mutex<HashMap<String, DiskAnalyzerSession>>>,
-        event_tx: broadcast::Sender<DiskAnalyzerScanEvent>,
-        scan_id: String,
-        owner: String,
-        level_path: PathBuf,
-        mut roots: DiskScanRoots,
-        max_children: usize,
-        cancel: Arc<AtomicBool>,
-        is_session_root: bool,
-    ) {
-        let engine = DiskAnalyzerEngine::new();
-        let level_key = path_key(&level_path);
-
-        let on_progress: ProgressCallback = Arc::new({
-            let event_tx = event_tx.clone();
-            let owner = owner.clone();
-            let sessions = Arc::clone(&sessions);
-            let scan_id = scan_id.clone();
-            move |progress: ScanProgress| {
-                if let Some(tree) = progress.tree.as_ref() {
-                    if let Some(session) = sessions.lock().get_mut(&scan_id) {
-                        if is_session_root
-                            || session.root_path.to_string_lossy().as_ref() == tree.path.as_str()
-                        {
-                            session.tree = Some(Arc::new(tree.clone()));
-                        } else if let Some(existing) = session.tree.as_ref() {
-                            let mut merged = (**existing).clone();
-                            if merge_subtree(&mut merged, tree) {
-                                session.tree = Some(Arc::new(merged));
-                            }
-                        } else {
-                            session.tree = Some(Arc::new(tree.clone()));
-                        }
-                    }
-                }
-                let Ok(payload) = serde_json::to_value(&progress) else {
-                    return;
-                };
-                let _ = event_tx.send(DiskAnalyzerScanEvent {
-                    owner_conn_id: owner.clone(),
-                    payload,
-                });
-            }
-        });
-
-        tokio::task::spawn_blocking(move || {
-            let need_discover = sessions
-                .lock()
-                .get(&scan_id)
-                .map(|s| !s.marks_ready)
-                .unwrap_or(true);
-            if need_discover {
-                if let Some(home) = dirs::home_dir() {
-                    roots.agent_data_roots = agent_data_roots(&home)
-                        .into_iter()
-                        .map(|(_, p)| p)
-                        .collect();
-                    roots.git_worktree_roots =
-                        GitEngine::new().discover_linked_worktrees_fast(&home, Some(&cancel));
-                }
-                if let Some(session) = sessions.lock().get_mut(&scan_id) {
-                    session.git_worktree_roots = roots.git_worktree_roots.clone();
-                    session.agent_data_roots = roots.agent_data_roots.clone();
-                    session.marks_ready = true;
-                }
-            } else if let Some(session) = sessions.lock().get(&scan_id) {
-                roots.git_worktree_roots = session.git_worktree_roots.clone();
-                roots.agent_data_roots = session.agent_data_roots.clone();
-            }
-            let result = engine.scan_path(
-                &scan_id,
-                &level_path,
-                &roots,
-                Some(max_children),
-                Some(cancel),
-                Some(on_progress),
-            );
-
-            match result {
-                Ok((tree, stats, suggestions)) => {
-                    let tree = Arc::new(tree);
-                    let payload = json!({
-                        "scan_id": scan_id,
-                        "status": if is_session_root { "completed" } else { "level_completed" },
-                        "files_scanned": stats.files_scanned,
-                        "bytes_scanned": stats.total_size,
-                        "dirs_scanned": stats.dirs_scanned,
-                        "error_count": stats.error_count,
-                        "current_path": Value::Null,
-                        "percent": 100.0,
-                        "error": Value::Null,
-                        "tree": tree.as_ref(),
-                        "level_path": level_key,
-                        "stats": &stats,
-                        "suggestions": if is_session_root { json!(&suggestions) } else { Value::Null },
-                    });
-                    if let Some(session) = sessions.lock().get_mut(&scan_id) {
-                        if is_session_root
-                            || session.root_path.to_string_lossy().as_ref() == tree.path.as_str()
-                        {
-                            session.tree = Some(Arc::clone(&tree));
-                        } else if let Some(existing) = session.tree.as_ref() {
-                            let mut merged = (**existing).clone();
-                            if merge_subtree(&mut merged, tree.as_ref()) {
-                                session.tree = Some(Arc::new(merged));
-                            } else {
-                                session.tree = Some(Arc::clone(&tree));
-                            }
-                        } else {
-                            session.tree = Some(Arc::clone(&tree));
-                        }
-                        session.inflight_root = None;
-                        session.stats = Some(stats);
-                        if is_session_root {
-                            session.suggestions = Some(suggestions);
-                            session.completed_at = Some(Instant::now());
-                        }
-                    }
-                    let _ = event_tx.send(DiskAnalyzerScanEvent {
-                        owner_conn_id: owner,
-                        payload,
-                    });
-                }
-                Err(e) => {
-                    let cancelled = e.to_string().contains("cancelled");
-                    let status = if cancelled {
-                        ScanStatus::Cancelled
-                    } else {
-                        ScanStatus::Failed
-                    };
-                    if let Some(session) = sessions.lock().get_mut(&scan_id) {
-                        session.inflight_root = None;
-                        if is_session_root {
-                            session.completed_at = Some(Instant::now());
-                        }
-                    }
-                    let payload = json!({
-                        "scan_id": scan_id,
-                        "status": status,
-                        "files_scanned": 0,
-                        "bytes_scanned": 0,
-                        "dirs_scanned": 0,
-                        "error_count": if cancelled { 0 } else { 1 },
-                        "current_path": Value::Null,
-                        "percent": Value::Null,
-                        "error": if cancelled { Value::Null } else { json!(e.to_string()) },
-                        "level_path": level_key,
-                    });
-                    let _ = event_tx.send(DiskAnalyzerScanEvent {
-                        owner_conn_id: owner,
-                        payload,
-                    });
-                }
-            }
-        });
-    }
-
     /// Returns `(project_roots, workspace_roots, project_labels)` where
     /// `project_labels` is `(display_name, path)` for overview tiles.
-    async fn collect_atmos_locations_static(
+    pub(super) async fn collect_atmos_locations_static(
         project_service: Arc<ProjectService>,
         workspace_service: Arc<WorkspaceService>,
         git_engine: GitEngine,
@@ -1749,7 +1088,7 @@ impl DiskAnalyzerService {
                 project_labels.push((project.name.clone(), root));
             }
             if let Ok(workspaces) = workspace_service
-                .list_by_project(project.guid.clone())
+                .list_all_by_project(project.guid.clone())
                 .await
             {
                 for workspace in workspaces {
@@ -1769,114 +1108,9 @@ impl DiskAnalyzerService {
         }
         (project_roots, workspace_roots, project_labels)
     }
-
-    pub fn remove_connection_sessions(&self, owner_conn_id: &str) {
-        let mut sessions = self.sessions.lock();
-        Self::cancel_and_purge_owner(&mut sessions, owner_conn_id);
-    }
-
-    fn ensure_owner(session: &DiskAnalyzerSession, owner_conn_id: &str) -> Result<()> {
-        if session.owner_conn_id != owner_conn_id {
-            return Err(ServiceError::Validation(
-                "scan session belongs to another connection".into(),
-            ));
-        }
-        Ok(())
-    }
-
-    fn ensure_under_any(path: &Path, roots: &[PathBuf]) -> Result<()> {
-        let path = std::fs::canonicalize(path).unwrap_or_else(|_| path.to_path_buf());
-        for root in roots {
-            let root = std::fs::canonicalize(root).unwrap_or_else(|_| root.clone());
-            if path == root || path.starts_with(&root) {
-                return Ok(());
-            }
-        }
-        Err(ServiceError::Validation(format!(
-            "path {} is outside scan entry roots",
-            path.display()
-        )))
-    }
-
-    /// Cancel in-flight walks and remove every session owned by `owner_conn_id`.
-    fn cancel_and_purge_owner(
-        sessions: &mut HashMap<String, DiskAnalyzerSession>,
-        owner_conn_id: &str,
-    ) {
-        sessions.retain(|_, session| {
-            if session.owner_conn_id == owner_conn_id {
-                session.cancel.store(true, Ordering::Relaxed);
-                false
-            } else {
-                true
-            }
-        });
-    }
-
-    fn evict_expired(sessions: &mut HashMap<String, DiskAnalyzerSession>) {
-        let now = Instant::now();
-        sessions.retain(|_, session| match session.completed_at {
-            Some(completed) => now.duration_since(completed) < SESSION_TTL,
-            None => {
-                if now.duration_since(session.started_at) >= SESSION_TTL * 2 {
-                    session.cancel.store(true, Ordering::Relaxed);
-                    false
-                } else {
-                    true
-                }
-            }
-        });
-    }
-
-    fn evict_oldest_completed(sessions: &mut HashMap<String, DiskAnalyzerSession>) {
-        let oldest = sessions
-            .iter()
-            .filter_map(|(id, s)| s.completed_at.map(|t| (id.clone(), t)))
-            .min_by_key(|(_, t)| *t);
-        if let Some((id, _)) = oldest {
-            sessions.remove(&id);
-        }
-    }
-
-    /// Evict the oldest session of any kind (completed preferred, else oldest start).
-    /// Returns false when the map is empty.
-    fn evict_oldest_any(sessions: &mut HashMap<String, DiskAnalyzerSession>) -> bool {
-        if sessions.is_empty() {
-            return false;
-        }
-        // Prefer completed sessions first.
-        if sessions.values().any(|s| s.completed_at.is_some()) {
-            Self::evict_oldest_completed(sessions);
-            return true;
-        }
-        let oldest = sessions
-            .iter()
-            .min_by_key(|(_, s)| s.started_at)
-            .map(|(id, _)| id.clone());
-        if let Some(id) = oldest {
-            if let Some(session) = sessions.remove(&id) {
-                session.cancel.store(true, Ordering::Relaxed);
-            }
-            true
-        } else {
-            false
-        }
-    }
 }
 
-fn path_key(path: &Path) -> String {
-    path.to_string_lossy().to_string()
-}
-
-#[derive(Debug, Clone, Copy)]
-enum OverviewKind {
-    /// Home tree with `/Applications` grafted under home.
-    HomeWithApps,
-    /// Synthetic "Atmos" root; each entry is a top-level child.
-    AtmosSynthetic,
-}
-
-fn synthetic_group(
+pub(super) fn synthetic_group(
     name: &str,
     path: &str,
     mut children: Vec<DiskNode>,
@@ -1904,7 +1138,7 @@ fn synthetic_group(
 }
 
 /// Merge parallel entry scans into one overview tree.
-fn assemble_overview(
+pub(super) fn assemble_overview(
     kind: OverviewKind,
     parts: &HashMap<String, DiskNode>,
     root_path: &str,
@@ -2011,584 +1245,5 @@ fn assemble_overview(
                 children,
             })
         }
-    }
-}
-
-fn find_node<'a>(node: &'a DiskNode, path: &str) -> Option<&'a DiskNode> {
-    if node.path == path {
-        return Some(node);
-    }
-    for child in &node.children {
-        if let Some(found) = find_node(child, path) {
-            return Some(found);
-        }
-    }
-    None
-}
-
-fn merge_subtree(root: &mut DiskNode, patch: &DiskNode) -> bool {
-    if root.path == patch.path {
-        *root = patch.clone();
-        return true;
-    }
-    for child in &mut root.children {
-        if merge_subtree(child, patch) {
-            if !root.children.is_empty() {
-                let child_sum: u64 = root.children.iter().map(|c| c.size).sum();
-                if child_sum > root.size {
-                    root.size = child_sum;
-                }
-            }
-            return true;
-        }
-    }
-    false
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn remove_connection_sessions_cancels_and_purges() {
-        let cancel = Arc::new(AtomicBool::new(false));
-        let mut sessions = HashMap::new();
-        sessions.insert(
-            "s1".to_string(),
-            DiskAnalyzerSession {
-                owner_conn_id: "conn-123".to_string(),
-                cancel: Arc::clone(&cancel),
-                tree: None,
-                inflight_root: None,
-                stats: None,
-                suggestions: None,
-                root_path: PathBuf::from("/tmp/home"),
-                entry_roots: vec![PathBuf::from("/tmp/home"), PathBuf::from("/Applications")],
-                max_children: 30,
-                project_roots: vec![],
-                workspace_roots: vec![],
-                git_worktree_roots: vec![],
-                agent_data_roots: vec![],
-                marks_ready: false,
-                started_at: Instant::now(),
-                completed_at: None,
-                is_overview: true,
-            },
-        );
-
-        let db = Arc::new(sea_orm::DatabaseConnection::default());
-        let service = DiskAnalyzerService {
-            engine: DiskAnalyzerEngine::new(),
-            fs_engine: FsEngine::new(),
-            project_service: Arc::new(ProjectService::new(Arc::clone(&db))),
-            workspace_service: Arc::new(WorkspaceService::new(Arc::clone(&db))),
-            sessions: Arc::new(Mutex::new(sessions)),
-            event_tx: broadcast::channel(16).0,
-        };
-
-        service.remove_connection_sessions("conn-123");
-        assert!(cancel.load(Ordering::Relaxed));
-        assert!(service.sessions.lock().is_empty());
-    }
-
-    #[test]
-    fn evict_expired_removes_old_in_flight_sessions() {
-        let mut sessions = HashMap::new();
-        let cancel = Arc::new(AtomicBool::new(false));
-        sessions.insert(
-            "s1".to_string(),
-            DiskAnalyzerSession {
-                owner_conn_id: "conn-1".to_string(),
-                cancel: Arc::clone(&cancel),
-                tree: None,
-                inflight_root: None,
-                stats: None,
-                suggestions: None,
-                root_path: PathBuf::from("/tmp"),
-                entry_roots: vec![PathBuf::from("/tmp")],
-                max_children: 30,
-                project_roots: vec![],
-                workspace_roots: vec![],
-                git_worktree_roots: vec![],
-                agent_data_roots: vec![],
-                marks_ready: false,
-                started_at: Instant::now() - (SESSION_TTL * 2 + Duration::from_secs(1)),
-                completed_at: None,
-                is_overview: false,
-            },
-        );
-
-        DiskAnalyzerService::evict_expired(&mut sessions);
-        assert!(cancel.load(Ordering::Relaxed));
-        assert!(sessions.is_empty());
-    }
-
-    #[test]
-    fn ensure_under_any_accepts_child() {
-        let tmp = std::env::temp_dir();
-        let nested = tmp.join("disk-analyzer-root-check");
-        let _ = std::fs::create_dir_all(&nested);
-        assert!(DiskAnalyzerService::ensure_under_any(&nested, &[tmp]).is_ok());
-    }
-
-    /// Drill-in must spawn a level walk even when session.tree already exists
-    /// (overview shells use children_loaded=false). Regression for the
-    /// `if inflight || cached { return loading }` short-circuit bug.
-    #[tokio::test]
-    async fn get_tree_spawns_walk_for_unloaded_directory() {
-        let base =
-            std::env::temp_dir().join(format!("disk-analyzer-get-tree-{}", std::process::id()));
-        let entry = base.join(".atmos");
-        let nested = entry.join("cache");
-        let _ = std::fs::remove_dir_all(&base);
-        std::fs::create_dir_all(&nested).expect("mkdir");
-        std::fs::write(nested.join("blob.bin"), vec![0u8; 4096]).expect("write");
-
-        let entry_key = path_key(&entry);
-        let overview = DiskNode {
-            name: "Atmos".into(),
-            path: ATMOS_OVERVIEW_PATH.into(),
-            size: 4096,
-            is_dir: true,
-            is_project: false,
-            is_workspace: false,
-            is_git_worktree: false,
-            is_agent_data: false,
-            file_count: 1,
-            dir_count: 1,
-            children_loaded: true,
-            children: vec![DiskNode {
-                name: ".atmos".into(),
-                path: entry_key.clone(),
-                size: 4096,
-                is_dir: true,
-                is_project: false,
-                is_workspace: false,
-                is_git_worktree: false,
-                is_agent_data: false,
-                file_count: 1,
-                dir_count: 1,
-                children_loaded: false,
-                children: vec![],
-            }],
-        };
-
-        let mut sessions = HashMap::new();
-        sessions.insert(
-            "scan-1".to_string(),
-            DiskAnalyzerSession {
-                owner_conn_id: "conn-1".to_string(),
-                cancel: Arc::new(AtomicBool::new(false)),
-                tree: Some(Arc::new(overview)),
-                inflight_root: None,
-                stats: None,
-                suggestions: None,
-                root_path: PathBuf::from(ATMOS_OVERVIEW_PATH),
-                entry_roots: vec![entry.clone()],
-                max_children: 30,
-                project_roots: vec![],
-                workspace_roots: vec![],
-                git_worktree_roots: vec![],
-                agent_data_roots: vec![],
-                marks_ready: false,
-                started_at: Instant::now(),
-                completed_at: Some(Instant::now()),
-                is_overview: true,
-            },
-        );
-
-        let db = Arc::new(sea_orm::DatabaseConnection::default());
-        let service = DiskAnalyzerService {
-            engine: DiskAnalyzerEngine::new(),
-            fs_engine: FsEngine::new(),
-            project_service: Arc::new(ProjectService::new(Arc::clone(&db))),
-            workspace_service: Arc::new(WorkspaceService::new(Arc::clone(&db))),
-            sessions: Arc::new(Mutex::new(sessions)),
-            event_tx: broadcast::channel(16).0,
-        };
-
-        let resp = service
-            .get_tree("conn-1", "scan-1", Some(&entry_key), Some(30))
-            .expect("get_tree");
-        assert_eq!(resp["status"], "loading");
-        assert_eq!(
-            service
-                .sessions
-                .lock()
-                .get("scan-1")
-                .unwrap()
-                .inflight_root
-                .as_deref(),
-            Some(entry_key.as_str()),
-            "unloaded dir must mark inflight so a walk is spawned"
-        );
-
-        let _ = std::fs::remove_dir_all(&base);
-    }
-
-    #[test]
-    fn get_tree_returns_ready_for_loaded_directory() {
-        let base = std::env::temp_dir().join(format!(
-            "disk-analyzer-get-tree-ready-{}",
-            std::process::id()
-        ));
-        let entry = base.join(".atmos");
-        let _ = std::fs::remove_dir_all(&base);
-        std::fs::create_dir_all(&entry).expect("mkdir");
-
-        let entry_key = path_key(&entry);
-        let overview = DiskNode {
-            name: "Atmos".into(),
-            path: ATMOS_OVERVIEW_PATH.into(),
-            size: 100,
-            is_dir: true,
-            is_project: false,
-            is_workspace: false,
-            is_git_worktree: false,
-            is_agent_data: false,
-            file_count: 1,
-            dir_count: 1,
-            children_loaded: true,
-            children: vec![DiskNode {
-                name: ".atmos".into(),
-                path: entry_key.clone(),
-                size: 100,
-                is_dir: true,
-                is_project: false,
-                is_workspace: false,
-                is_git_worktree: false,
-                is_agent_data: false,
-                file_count: 1,
-                dir_count: 0,
-                children_loaded: true,
-                children: vec![DiskNode {
-                    name: "blob.bin".into(),
-                    path: format!("{entry_key}/blob.bin"),
-                    size: 100,
-                    is_dir: false,
-                    is_project: false,
-                    is_workspace: false,
-                    is_git_worktree: false,
-                    is_agent_data: false,
-                    file_count: 1,
-                    dir_count: 0,
-                    children_loaded: true,
-                    children: vec![],
-                }],
-            }],
-        };
-
-        let mut sessions = HashMap::new();
-        sessions.insert(
-            "scan-1".to_string(),
-            DiskAnalyzerSession {
-                owner_conn_id: "conn-1".to_string(),
-                cancel: Arc::new(AtomicBool::new(false)),
-                tree: Some(Arc::new(overview)),
-                inflight_root: None,
-                stats: None,
-                suggestions: None,
-                root_path: PathBuf::from(ATMOS_OVERVIEW_PATH),
-                entry_roots: vec![entry.clone()],
-                max_children: 30,
-                project_roots: vec![],
-                workspace_roots: vec![],
-                git_worktree_roots: vec![],
-                agent_data_roots: vec![],
-                marks_ready: false,
-                started_at: Instant::now(),
-                completed_at: Some(Instant::now()),
-                is_overview: true,
-            },
-        );
-
-        let db = Arc::new(sea_orm::DatabaseConnection::default());
-        let service = DiskAnalyzerService {
-            engine: DiskAnalyzerEngine::new(),
-            fs_engine: FsEngine::new(),
-            project_service: Arc::new(ProjectService::new(Arc::clone(&db))),
-            workspace_service: Arc::new(WorkspaceService::new(Arc::clone(&db))),
-            sessions: Arc::new(Mutex::new(sessions)),
-            event_tx: broadcast::channel(16).0,
-        };
-
-        let resp = service
-            .get_tree("conn-1", "scan-1", Some(&entry_key), Some(30))
-            .expect("get_tree");
-        assert_eq!(resp["status"], "ready");
-        assert_eq!(resp["tree"]["name"], ".atmos");
-        assert!(
-            resp["tree"]["children"]
-                .as_array()
-                .map(|a| !a.is_empty())
-                .unwrap_or(false),
-            "loaded dir should include children"
-        );
-
-        let _ = std::fs::remove_dir_all(&base);
-    }
-
-    fn test_leaf(name: &str, path: &str, size: u64) -> DiskNode {
-        DiskNode {
-            name: name.into(),
-            path: path.into(),
-            size,
-            is_dir: false,
-            is_project: false,
-            is_workspace: false,
-            is_git_worktree: false,
-            is_agent_data: false,
-            file_count: 1,
-            dir_count: 0,
-            children_loaded: true,
-            children: vec![],
-        }
-    }
-
-    #[tokio::test]
-    async fn get_tree_rewalks_when_requested_max_exceeds_pruned_cache() {
-        let base = std::env::temp_dir().join(format!(
-            "disk-analyzer-get-tree-wider-{}",
-            std::process::id()
-        ));
-        let entry = base.join(".atmos");
-        let _ = std::fs::remove_dir_all(&base);
-        std::fs::create_dir_all(&entry).expect("mkdir");
-        for i in 0..5 {
-            std::fs::write(entry.join(format!("f{i}.bin")), vec![0u8; 64]).expect("write");
-        }
-
-        let entry_key = path_key(&entry);
-        let pruned = DiskNode {
-            name: ".atmos".into(),
-            path: entry_key.clone(),
-            size: 320,
-            is_dir: true,
-            is_project: false,
-            is_workspace: false,
-            is_git_worktree: false,
-            is_agent_data: false,
-            file_count: 5,
-            dir_count: 0,
-            children_loaded: true,
-            children: vec![
-                test_leaf("f0.bin", &format!("{entry_key}/f0.bin"), 64),
-                test_leaf("f1.bin", &format!("{entry_key}/f1.bin"), 64),
-                DiskNode {
-                    name: "__other__".into(),
-                    path: format!("{entry_key}/__other__"),
-                    size: 192,
-                    is_dir: true,
-                    is_project: false,
-                    is_workspace: false,
-                    is_git_worktree: false,
-                    is_agent_data: false,
-                    file_count: 3,
-                    dir_count: 0,
-                    children_loaded: true,
-                    children: vec![],
-                },
-            ],
-        };
-
-        let mut sessions = HashMap::new();
-        sessions.insert(
-            "scan-1".to_string(),
-            DiskAnalyzerSession {
-                owner_conn_id: "conn-1".to_string(),
-                cancel: Arc::new(AtomicBool::new(false)),
-                tree: Some(Arc::new(pruned)),
-                inflight_root: None,
-                stats: None,
-                suggestions: None,
-                root_path: PathBuf::from(ATMOS_OVERVIEW_PATH),
-                entry_roots: vec![entry.clone()],
-                max_children: 2,
-                project_roots: vec![],
-                workspace_roots: vec![],
-                git_worktree_roots: vec![],
-                agent_data_roots: vec![],
-                marks_ready: false,
-                started_at: Instant::now(),
-                completed_at: Some(Instant::now()),
-                is_overview: true,
-            },
-        );
-
-        let db = Arc::new(sea_orm::DatabaseConnection::default());
-        let service = DiskAnalyzerService {
-            engine: DiskAnalyzerEngine::new(),
-            fs_engine: FsEngine::new(),
-            project_service: Arc::new(ProjectService::new(Arc::clone(&db))),
-            workspace_service: Arc::new(WorkspaceService::new(Arc::clone(&db))),
-            sessions: Arc::new(Mutex::new(sessions)),
-            event_tx: broadcast::channel(16).0,
-        };
-
-        let same_cap = service
-            .get_tree("conn-1", "scan-1", Some(&entry_key), Some(2))
-            .expect("get_tree same cap");
-        assert_eq!(same_cap["status"], "ready");
-        assert_eq!(
-            service
-                .sessions
-                .lock()
-                .get("scan-1")
-                .unwrap()
-                .inflight_root
-                .as_deref(),
-            None,
-            "same cap should serve the pruned snapshot"
-        );
-
-        let wider = service
-            .get_tree("conn-1", "scan-1", Some(&entry_key), Some(5))
-            .expect("get_tree wider");
-        assert_eq!(wider["status"], "loading");
-        assert_eq!(
-            service
-                .sessions
-                .lock()
-                .get("scan-1")
-                .unwrap()
-                .inflight_root
-                .as_deref(),
-            Some(entry_key.as_str()),
-            "wider cap must re-walk instead of returning the pruned snapshot"
-        );
-        assert_eq!(
-            service.sessions.lock().get("scan-1").unwrap().max_children,
-            5
-        );
-
-        let _ = std::fs::remove_dir_all(&base);
-    }
-
-    #[test]
-    fn append_discovered_skips_covered_paths_and_labels_worktrees() {
-        let tmp = std::env::temp_dir().join(format!(
-            "disk-analyzer-discover-entries-{}",
-            std::process::id()
-        ));
-        let _ = std::fs::remove_dir_all(&tmp);
-        let atmos = tmp.join(".atmos");
-        let cursor_sessions = tmp.join(".cursor").join("projects");
-        let nested_wt = tmp.join(".cursor").join("worktrees").join("feat");
-        let extra_wt = tmp.join("extra-wt");
-        std::fs::create_dir_all(&atmos).expect("atmos");
-        std::fs::create_dir_all(&cursor_sessions).expect("cursor sessions");
-        std::fs::create_dir_all(&nested_wt).expect("nested wt");
-        std::fs::create_dir_all(&extra_wt).expect("extra wt");
-
-        let mut entries = vec![EntryRoot {
-            label: ".atmos".into(),
-            path: atmos,
-            kind: None,
-        }];
-        DiskAnalyzerService::append_discovered_overview_entries(
-            &mut entries,
-            &[nested_wt.clone(), extra_wt.clone()],
-            &[("cursor".into(), cursor_sessions)],
-        );
-
-        let labels: Vec<_> = entries.iter().map(|e| e.label.as_str()).collect();
-        assert!(labels.contains(&".atmos"));
-        assert!(labels.contains(&"cursor"));
-        assert!(
-            entries
-                .iter()
-                .any(|e| e.path == extra_wt && e.kind == Some(AtmosEntryKind::GitWorktree)),
-            "uncovered worktree should become an overview tile: {labels:?}"
-        );
-        assert!(
-            entries
-                .iter()
-                .any(|e| e.path == nested_wt && e.kind == Some(AtmosEntryKind::GitWorktree)),
-            "worktree under .cursor/worktrees is not covered by session dir: {labels:?}"
-        );
-        assert!(
-            DiskAnalyzerService::worktree_overview_label(&extra_wt).starts_with("extra-wt"),
-            "worktree label should start with the directory name"
-        );
-
-        let _ = std::fs::remove_dir_all(&tmp);
-    }
-
-    fn overview_part(
-        name: &str,
-        path: &str,
-        size: u64,
-        is_git_worktree: bool,
-        is_agent_data: bool,
-    ) -> DiskNode {
-        DiskNode {
-            name: name.into(),
-            path: path.into(),
-            size,
-            is_dir: true,
-            is_project: false,
-            is_workspace: false,
-            is_git_worktree,
-            is_agent_data,
-            file_count: 1,
-            dir_count: 1,
-            children_loaded: false,
-            children: vec![],
-        }
-    }
-
-    #[test]
-    fn assemble_overview_groups_agent_and_worktree_entries() {
-        let mut parts = HashMap::new();
-        parts.insert(
-            ".atmos".into(),
-            overview_part(".atmos", "/home/u/.atmos", 100, false, false),
-        );
-        parts.insert(
-            "cursor".into(),
-            overview_part("cursor", "/home/u/.cursor/projects", 50, false, true),
-        );
-        parts.insert(
-            "feat".into(),
-            overview_part("feat", "/tmp/feat", 80, true, false),
-        );
-
-        let tree = assemble_overview(
-            OverviewKind::AtmosSynthetic,
-            &parts,
-            ATMOS_OVERVIEW_PATH,
-            ATMOS_OVERVIEW_NAME,
-            &HashMap::new(),
-        )
-        .expect("overview");
-
-        let names: Vec<_> = tree.children.iter().map(|c| c.name.as_str()).collect();
-        assert!(
-            names.contains(&".atmos"),
-            "Atmos runtime stays at root: {names:?}"
-        );
-        assert!(
-            !names.contains(&"cursor") && !names.contains(&"feat"),
-            "agent/worktree tiles must nest under groups: {names:?}"
-        );
-
-        let agent = tree
-            .children
-            .iter()
-            .find(|c| c.path == AGENT_GROUP_PATH)
-            .expect("agent group");
-        assert!(agent.is_agent_data);
-        assert!(agent.children_loaded);
-        assert!(agent.children.iter().any(|c| c.name == "cursor"));
-
-        let worktrees = tree
-            .children
-            .iter()
-            .find(|c| c.path == WORKTREE_GROUP_PATH)
-            .expect("worktree group");
-        assert!(worktrees.is_git_worktree);
-        assert!(worktrees.children_loaded);
-        assert!(worktrees.children.iter().any(|c| c.name == "feat"));
-        assert_eq!(tree.size, 230);
     }
 }
